refactor(product): move files to Domain/Contracts/Abstractions (Phase 2)
- Product/ProductType -> Domain/, IProductStore -> Domain/Abstractions/,
  integration events -> Contracts/Integration/; namespaces renamed.
- Types stay anemic; HTTP routes, event payloads, metrics unchanged.
- EF model-snapshot entity names updated to the new Domain namespace.
