+++ product-microservice/Product.Service/Contracts/Integration/ProductCreatedEvent.cs
@@ -1,5 +1,5 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Product.Service.IntegrationEvents;
+namespace Product.Service.Contracts.Integration;
 
 public record ProductCreatedEvent(int ProductId, string Name, decimal Price) : Event;
