+++ product-microservice/Product.Service/Contracts/Integration/ProductPriceUpdatedEvent.cs
@@ -1,5 +1,5 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Product.Service.IntegrationEvents;
+namespace Product.Service.Contracts.Integration;
 
 public record ProductPriceUpdatedEvent(int ProductId, decimal NewPrice) : Event;
