+++ product-microservice/Product.Service/Domain/Abstractions/IProductStore.cs
@@ -1,0 +1,10 @@
+namespace Product.Service.Domain.Abstractions;
+
+internal interface IProductStore
+{
+    Task<Product?> GetById(int id);
+
+    Task CreateProduct(Product product);
+
+    Task UpdateProduct(Product product);
+}
