+++ product-microservice/Product.Service/Domain/Product.cs
@@ -1,4 +1,4 @@
-namespace Product.Service.Models;
+namespace Product.Service.Domain;
 
 internal class Product
 {
