+++ product-microservice/Product.Service/Infrastructure/Data/EntityFramework/ProductContextSeed.cs
@@ -1,5 +1,4 @@
 using Microsoft.EntityFrameworkCore;
-using Product.Service.Models;
 
 namespace Product.Service.Infrastructure.Data.EntityFramework;
